Add tests for model_utils and fix install script issue
